Package structure: moved Marker UI code.
Part of #963.
