+++ src/main/java/de/dennisguse/opentracks/chart/ChartView.java
@@ -44,12 +44,12 @@ import java.util.ArrayList;
 import java.util.LinkedList;
 import java.util.List;
 
-import de.dennisguse.opentracks.MarkerDetailActivity;
 import de.dennisguse.opentracks.R;
 import de.dennisguse.opentracks.content.data.Marker;
 import de.dennisguse.opentracks.stats.ExtremityMonitor;
+import de.dennisguse.opentracks.ui.markers.MarkerDetailActivity;
+import de.dennisguse.opentracks.ui.markers.MarkerUtils;
 import de.dennisguse.opentracks.util.IntentUtils;
-import de.dennisguse.opentracks.util.MarkerUtils;
 import de.dennisguse.opentracks.util.StringUtils;
 import de.dennisguse.opentracks.util.ThemeUtils;
