+++ src/main/java/de/dennisguse/opentracks/ui/markers/MarkerEditViewModel.java
@@ -1,4 +1,4 @@
-package de.dennisguse.opentracks.viewmodels;
+package de.dennisguse.opentracks.ui.markers;
 
 import android.app.Application;
 import android.net.Uri;
